requirements.txt
cryptography==41.0.4
pyjwt==2.8.0
requests==2.31.0
pytest==7.4.3
pytest-cov==4.1.0
